Move the `hooks-server` module under the `@bayou` namespace.
Also, extract a dependency.
