+++ local-modules/@bayou/deps-util-server/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "async-file": "^2.0.2",
     "chokidar": "^1.6.1",
+    "content-type": "^1.0.2",
     "fs-extra": "^3.0.1",
     "is-running": "^2.1.0",
     "memory-fs": "^0.4.1",
