+++ local-modules/@bayou/hooks-server/default-document.js
