+++ local-modules/@bayou/hooks-server/tests/test_DEFAULT_DOCUMENT.js
@@ -5,11 +5,11 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { DEFAULT_DOCUMENT } from 'hooks-server';
+import { DEFAULT_DOCUMENT } from '@bayou/hooks-server';
 import { BodyDelta } from '@bayou/doc-common';
 import { DataUtil } from '@bayou/util-common';
 
-describe('hooks-server/DEFAULT_DOCUMENT', () => {
+describe('@bayou/hooks-server/DEFAULT_DOCUMENT', () => {
   it('should be a deep-frozen data value', () => {
     const doc = DEFAULT_DOCUMENT;
