+++ local-modules/api-server/package.json
@@ -3,11 +3,10 @@
   "version": "1.0.0",
 
   "dependencies": {
-    "content-type": "^1.0.2",
-
     "@bayou/api-common": "local",
     "@bayou/codec": "local",
-    "hooks-server": "local",
+    "@bayou/deps-util-server": "local",
+    "@bayou/hooks-server": "local",
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/typecheck": "local",
