+++ local-modules/doc-server/DocServer.js
@@ -5,7 +5,7 @@
 import weak from 'weak';
 
 import { TheModule as appCommon_TheModule } from '@bayou/app-common';
-import { Hooks } from 'hooks-server';
+import { Hooks } from '@bayou/hooks-server';
 import { Logger } from '@bayou/see-all';
 import { TFunction, TString } from '@bayou/typecheck';
 import { Singleton } from '@bayou/util-common';
