+++ local-modules/doc-server/package.json
@@ -11,7 +11,7 @@
     "env-server": "local",
     "file-store": "local",
     "@bayou/file-store-ot": "local",
-    "hooks-server": "local",
+    "@bayou/hooks-server": "local",
     "@bayou/ot-common": "local",
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
