+++ server/package.json
@@ -21,7 +21,7 @@
     "@bayou/deps-util-server": "local",
     "dev-mode": "local",
     "env-server": "local",
-    "hooks-server": "local",
+    "@bayou/hooks-server": "local",
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/see-all-server": "local",
